+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 ### Fixed
 - Fixed invalid import from "legacy" feature flag from stabilized collection. [PR 960](https://github.com/near/near-sdk-rs/pull/960)
 
+### Removed
+- Deprecated declarative macros for NFT impl code generation. [PR 1042](https://github.com/near/near-sdk-rs/pull/1042)
+
 ## [4.1.0] - 2022-11-09
 
 ### Added